requirements.txt
Django==4.2.28
Pillow>=10.0,<12.0
python-dotenv>=1.0,<2.0
gunicorn>=22.0,<24.0
whitenoise>=6.7,<7.0
dj-database-url>=2.2,<3.0
psycopg2-binary>=2.9,<3.0
